Better quant Bench: 1629737

+++ src/networks/accumulator.rs
@@ -77,7 +77,7 @@ impl<const N: usize> Accumulator<f32, N> {
         res
     }
 
-    pub fn quantise_i8(&self, qa: i8) -> Accumulator<i8, N> {
+    pub fn quantise_i8(&self, qa: i16) -> Accumulator<i8, N> {
         let mut res = Accumulator([0; N]);
 
         for (i, &j) in res.0.iter_mut().zip(self.0.iter()) {
